requirements.txt
pytest
pytest-html
requests
jsonschema
pyyaml
google-generativeai
